Punybench build process cleanups
Punybench build process will attempt to include a Makefile snippet for
the defined target architecture, and it will fail if the architecture
file is not present.

To prevent the error, the build system will now only include the file
if it exists; I also removed two architecture files which did nothing,
but were present so they could be included.

Finally, I made the 'makedir' variable hold an absolute pathname
instead of a relative pathname; if the build process fails, it will
show an absolute pathname now.

BUG=none
TEST=Inspection, and tested building with a non-existant architecture file.

Change-Id: I8dc3449b310677f88347980394d1abfaeaf0f785
Signed-off-by: Taylor Hutt <thutt@chromium.org>