Added testsync to help test sync.
Testsync works in conjunction with platform_SyncCrash to
test fsync and msync. The test writes data to a file,
calls a flavor of sync and then the crashes the system.
After the system comes back up, testsync verifies that
the data was saved correctly.

BUG=chromium:239536
TEST=platform_SyncCrash

Change-Id: Ieceba236a00b344ca96fdc5f44c52985e1c51492
Reviewed-on: https://gerrit.chromium.org/gerrit/56412
Tested-by: Paul Taysom <taysom@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Paul Taysom <taysom@chromium.org>