punybench: add option to skip trivial test
Much of the time, we're trying to test hardware limits on the
memory/cache interface and the trivial tests tend to run into core
limitations.  So it would be convienient to be able to run the tests
more quickly by skipping the trivial part.

BUG=chrome-os-partner:19854
TEST=`memcpy_test -bmt` runs and skips the trival tests
TEST=`memcpy_test -bm` still runs properly

Change-Id: Idbcba475b76e98643b631fd34ff3ca9fa5348989
Reviewed-on: https://gerrit.chromium.org/gerrit/62805
Reviewed-by: Paul Taysom <taysom@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>